+++ src/MonkeyPatch/FileStreamWrapper.php
@@ -12,7 +12,7 @@ class FileStreamWrapper
     protected static bool $isRegistered = false;
 
     /**
-     * @var array<string,PatcherInterface>
+     * @var array<string,array{0: PatcherInterface, 1: bool, 2: string}>
      */
     private static array $fileToPatcherMap = [];
 
@@ -22,14 +22,18 @@ class FileStreamWrapper
     public $context;
 
     /**
-     * @var resource
+     * @var resource|null
      */
-    private $fileResource;
+    private $fileResource = null;
 
-    public static function setPatcherForFile(string $file, PatcherInterface $patcher): void
-    {
+    public static function setPatcherForFile(
+        string $file,
+        PatcherInterface $patcher,
+        bool $redirectOpenedPath = true,
+        string $context = null
+    ): void {
         $fromFilePath = FS::realpath($file) ?: $file;
-        self::$fileToPatcherMap[$fromFilePath] = $patcher;
+        self::$fileToPatcherMap[$fromFilePath] = [$patcher, $redirectOpenedPath, $context ?? ''];
         self::register();
     }
 
@@ -116,6 +120,10 @@ class FileStreamWrapper
      */
     public function stream_cast()
     {
+        if (!is_resource($this->fileResource)) {
+            throw new MonkeyPatchingException('Cannot cast a non-resource to a resource.');
+        }
+
         return $this->fileResource;
     }
 
@@ -128,6 +136,10 @@ class FileStreamWrapper
 
     public function stream_set_option(int $option, int $arg1, int $arg2): bool
     {
+        if (!is_resource($this->fileResource)) {
+            return false;
+        }
+
         switch ($option) {
             case STREAM_OPTION_BLOCKING:
                 return stream_set_blocking($this->fileResource, (bool)$arg1);
@@ -146,6 +158,10 @@ class FileStreamWrapper
      */
     public function stream_write(string $data): int
     {
+        if (!is_resource($this->fileResource)) {
+            throw new MonkeyPatchingException('Cannot write to a non-resource.');
+        }
+
         $written = fwrite($this->fileResource, $data);
 
         if ($written === false) {
@@ -160,7 +176,7 @@ class FileStreamWrapper
      */
     public function stream_tell(): int
     {
-        $pos = ftell($this->fileResource);
+        $pos = is_resource($this->fileResource) ? ftell($this->fileResource) : false;
 
         if ($pos === false) {
             throw new MonkeyPatchingException('Could not get the position of the file pointer.');
@@ -171,16 +187,28 @@ class FileStreamWrapper
 
     public function stream_close(): bool
     {
+        if (!is_resource($this->fileResource)) {
+            return false;
+        }
+
         return fclose($this->fileResource);
     }
 
     public function stream_eof(): bool
     {
+        if (!is_resource($this->fileResource)) {
+            return true;
+        }
+
         return feof($this->fileResource);
     }
 
     public function stream_seek(int $offset, int $whence = SEEK_SET): bool
     {
+        if (!is_resource($this->fileResource)) {
+            return false;
+        }
+
         return fseek($this->fileResource, $offset, $whence) === 0;
     }
 
@@ -232,6 +260,10 @@ class FileStreamWrapper
      */
     public function stream_stat(): array|false
     {
+        if (!is_resource($this->fileResource)) {
+            return false;
+        }
+
         $stat = fstat($this->fileResource);
 
         if ($stat === false) {
@@ -304,6 +336,10 @@ class FileStreamWrapper
 
     public function stream_truncate(int $newSize): bool
     {
+        if (!is_resource($this->fileResource)) {
+            return false;
+        }
+
         return ftruncate($this->fileResource, max(0, $newSize));
     }
 
@@ -360,7 +396,7 @@ class FileStreamWrapper
         if (!(file_exists($path))) {
             if (isset(self::$fileToPatcherMap[$path])) {
                 // Ask the patcher to provide stats.
-                $stat = self::$fileToPatcherMap[$path]->stat($path);
+                $stat = self::$fileToPatcherMap[$path][0]->stat($path);
             } else {
                 $stat = false;
             }
@@ -437,7 +473,7 @@ class FileStreamWrapper
 
         static::register();
 
-        return $this->fileResource !== null;
+        return true;
     }
 
     /**
@@ -445,7 +481,7 @@ class FileStreamWrapper
      */
     private function patchFile(string $absPath): string
     {
-        $patcher = self::$fileToPatcherMap[$absPath];
+        [$patcher, $redirectOpenedPath, $context] = self::$fileToPatcherMap[$absPath];
         self::unregister();
         // Do not use `is_file` here as it will use the cached stats: this check should be real.
         $fileContents = file_exists($absPath) ? file_get_contents($absPath) : '';
@@ -455,7 +491,7 @@ class FileStreamWrapper
             throw new MonkeyPatchingException("Could not read file $absPath contents.");
         }
 
-        [$fileContents, $openedPath] = $patcher->patch($fileContents, $absPath);
+        [$fileContents, $openedPath] = $patcher->patch($fileContents, $absPath, $context);
 
         if ($this->context !== null) {
             $fileResource = fopen('php://temp', 'rb+', false, $this->context);
@@ -475,7 +511,7 @@ class FileStreamWrapper
 
         rewind($this->fileResource);
 
-        return $openedPath;
+        return $redirectOpenedPath ? $openedPath : $absPath;
     }
 
     /**