+++ src/TestCase/WPUnitTestCasePolyfillsTrait.php
@@ -9,7 +9,7 @@ trait WPUnitTestCasePolyfillsTrait
     /**
      * @param array<mixed> $arguments
      */
-    public function __call(string $name, array $arguments)
+    public function __call(string $name, array $arguments) //@phpstan-ignore-line cannot be type-hinted
     {
         return TestCase::$name(...$arguments);
     }
@@ -17,7 +17,7 @@ trait WPUnitTestCasePolyfillsTrait
     /**
      * @param array<mixed> $arguments
      */
-    public static function __callStatic(string $name, array $arguments)
+    public static function __callStatic(string $name, array $arguments) //@phpstan-ignore-line cannot be type-hinted
     {
         return TestCase::$name(...$arguments);
     }